requirements.txt
pillow==10.4.0
chromadb==0.5.5
openai==1.39.0
pdfplumber==0.11.3
requests==2.32.3
streamlit==1.37.1
tiktoken==0.7.0
torch==2.4.0
transformers==4.43.4
